smoosh debug stuff, remove extra group
