+++ apps/dotcom/client/src/tla/components/TlaAppMenu/TlaAppMenuContent.tsx
@@ -1,14 +1,18 @@
 import { useAuth } from '@clerk/clerk-react'
 import { fileOpen } from 'browser-fs-access'
+import { useCallback, useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import {
 	ColorSchemeMenu,
 	LanguageMenu,
 	TLDRAW_FILE_EXTENSION,
+	TldrawUiMenuCheckboxItem,
 	TldrawUiMenuGroup,
 	TldrawUiMenuItem,
+	TldrawUiMenuSubmenu,
 	useDialogs,
 	useMaybeEditor,
+	useValue,
 } from 'tldraw'
 import { useOpenUrlAndTrack } from '../../../hooks/useOpenUrlAndTrack'
 import { routes } from '../../../routeDefs'
@@ -16,7 +20,7 @@ import { useMaybeApp } from '../../hooks/useAppState'
 import { useTldrawAppUiEvents } from '../../utils/app-ui-events'
 import { getCurrentEditor } from '../../utils/getCurrentEditor'
 import { defineMessages, useMsg } from '../../utils/i18n'
-import { TlaDebugMenuGroup } from '../TlaDebugMenuGroup'
+import { clearLocalSessionState } from '../../utils/local-session-state'
 import { SubmitFeedbackDialog } from '../dialogs/SubmitFeedbackDialog'
 import { TlaManageCookiesDialog } from '../dialogs/TlaManageCookiesDialog'
 
@@ -31,9 +35,15 @@ const messages = defineMessages({
 	manageCookies: { defaultMessage: 'Manage cookies' },
 	about: { defaultMessage: 'About tldraw' },
 	submitFeedback: { defaultMessage: 'Send feedback' },
+	signOut: { defaultMessage: 'Sign out' },
+	// debug
+	appDebugFlags: { defaultMessage: 'App debug flags' },
+	langAccented: { defaultMessage: 'i18n: Accented' },
+	langLongString: { defaultMessage: 'i18n: Long String' },
+	langHighlightMissing: { defaultMessage: 'i18n: Highlight Missing' },
 })
 
-export function TlaAppMenuGroup() {
+export function TlaAppMenuContent() {
 	const isSignedIn = useAuth().isSignedIn
 
 	return (
@@ -49,10 +59,33 @@ export function TlaAppMenuGroup() {
 			<TldrawUiMenuGroup id="settings-and-rare-actions">
 				<ColorThemeSubmenu />
 				<LanguageMenu />
-				<TlaDebugMenuGroup />
+				<DebugMenuGroup />
 				<ImportFileActionItem />
 			</TldrawUiMenuGroup>
+			<TldrawUiMenuGroup id="signout">
+				<SignOutMenuItem />
+			</TldrawUiMenuGroup>
 		</>
+	)
+}
+
+function SignOutMenuItem() {
+	const auth = useAuth()
+
+	const trackEvent = useTldrawAppUiEvents()
+
+	const label = useMsg(messages.signOut)
+
+	const handleSignout = useCallback(() => {
+		auth.signOut().then(clearLocalSessionState)
+		trackEvent('sign-out-clicked', { source: 'sidebar' })
+	}, [auth, trackEvent])
+
+	if (!auth.isSignedIn) return
+	return (
+		<TldrawUiMenuGroup id="account-actions">
+			<TldrawUiMenuItem id="sign-out" label={label} readonlyOk onSelect={handleSignout} />
+		</TldrawUiMenuGroup>
 	)
 }
 
@@ -176,5 +209,62 @@ function ImportFileActionItem() {
 				}
 			}}
 		/>
+	)
+}
+
+function DebugMenuGroup() {
+	const maybeEditor = useMaybeEditor()
+	const isDebugMode = useValue('debug', () => maybeEditor?.getInstanceState().isDebugMode, [
+		maybeEditor,
+	])
+	if (!isDebugMode) return null
+
+	return <DebugSubmenu />
+}
+
+function DebugSubmenu() {
+	const editor = useMaybeEditor()
+	const appFlagsLbl = useMsg(messages.appDebugFlags)
+
+	const [shouldHighlightMissing, setShouldHighlightMissing] = useState(
+		document.body.classList.contains('tla-lang-highlight-missing')
+	)
+	const debugLanguageFlags = [
+		{ name: useMsg(messages.langAccented), locale: 'xx-AE' },
+		{ name: useMsg(messages.langLongString), locale: 'xx-LS' },
+		{ name: useMsg(messages.langHighlightMissing), locale: 'xx-MS' },
+	]
+
+	useEffect(() => {
+		document.body.classList.toggle('tla-lang-highlight-missing', shouldHighlightMissing)
+	}, [shouldHighlightMissing])
+
+	return (
+		<TldrawUiMenuSubmenu id="debug" label={appFlagsLbl}>
+			<TldrawUiMenuGroup id="debug app flags">
+				{debugLanguageFlags.map((flag) => (
+					<TldrawUiMenuCheckboxItem
+						key={flag.name}
+						id={flag.name}
+						title={flag.name}
+						label={flag.name
+							.replace(/([a-z0-9])([A-Z])/g, (m) => `${m[0]} ${m[1].toLowerCase()}`)
+							.replace(/^[a-z]/, (m) => m.toUpperCase())}
+						checked={
+							flag.locale === 'xx-MS'
+								? shouldHighlightMissing
+								: editor?.user.getLocale() === flag.locale
+						}
+						onSelect={() => {
+							if (flag.locale === 'xx-MS') {
+								setShouldHighlightMissing(!shouldHighlightMissing)
+							} else {
+								editor?.user.updateUserPreferences({ locale: flag.locale })
+							}
+						}}
+					/>
+				))}
+			</TldrawUiMenuGroup>
+		</TldrawUiMenuSubmenu>
 	)
 }
