--- apps/dotcom/client/src/tla/components/TlaDebugMenuGroup.tsx
@@ -1,67 +1,0 @@
-import { useEffect, useState } from 'react'
-import {
-	TldrawUiMenuCheckboxItem,
-	TldrawUiMenuGroup,
-	TldrawUiMenuSubmenu,
-	useMaybeEditor,
-	useValue,
-} from 'tldraw'
-import { useMsg } from '../utils/i18n'
-import { messages } from './TlaSidebar/components/sidebar-shared'
-
-export function TlaDebugMenuGroup() {
-	const maybeEditor = useMaybeEditor()
-	const isDebugMode = useValue('debug', () => maybeEditor?.getInstanceState().isDebugMode, [
-		maybeEditor,
-	])
-	if (!isDebugMode) return null
-
-	return <DebugSubmenu />
-}
-
-function DebugSubmenu() {
-	const editor = useMaybeEditor()
-	const appFlagsLbl = useMsg(messages.appDebugFlags)
-
-	const [shouldHighlightMissing, setShouldHighlightMissing] = useState(
-		document.body.classList.contains('tla-lang-highlight-missing')
-	)
-	const debugLanguageFlags = [
-		{ name: useMsg(messages.langAccented), locale: 'xx-AE' },
-		{ name: useMsg(messages.langLongString), locale: 'xx-LS' },
-		{ name: useMsg(messages.langHighlightMissing), locale: 'xx-MS' },
-	]
-
-	useEffect(() => {
-		document.body.classList.toggle('tla-lang-highlight-missing', shouldHighlightMissing)
-	}, [shouldHighlightMissing])
-
-	return (
-		<TldrawUiMenuSubmenu id="debug" label={appFlagsLbl}>
-			<TldrawUiMenuGroup id="debug app flags">
-				{debugLanguageFlags.map((flag) => (
-					<TldrawUiMenuCheckboxItem
-						key={flag.name}
-						id={flag.name}
-						title={flag.name}
-						label={flag.name
-							.replace(/([a-z0-9])([A-Z])/g, (m) => `${m[0]} ${m[1].toLowerCase()}`)
-							.replace(/^[a-z]/, (m) => m.toUpperCase())}
-						checked={
-							flag.locale === 'xx-MS'
-								? shouldHighlightMissing
-								: editor?.user.getLocale() === flag.locale
-						}
-						onSelect={() => {
-							if (flag.locale === 'xx-MS') {
-								setShouldHighlightMissing(!shouldHighlightMissing)
-							} else {
-								editor?.user.updateUserPreferences({ locale: flag.locale })
-							}
-						}}
-					/>
-				))}
-			</TldrawUiMenuGroup>
-		</TldrawUiMenuSubmenu>
-	)
-}
