+++ apps/dotcom/client/src/tla/components/TlaEditor/editor-messages.ts
@@ -3,7 +3,7 @@ import { defineMessages } from '../../utils/i18n'
 export const editorMessages = defineMessages({
 	file: { defaultMessage: 'File' },
 	untitledProject: { defaultMessage: 'Untitled file' },
-	downloadFile: { defaultMessage: 'Download file' },
+	downloadFile: { defaultMessage: 'Download' },
 	copyToMyfiles: { defaultMessage: 'Copy to my files' },
 	anonymousUser: { defaultMessage: 'Guest user' },
 })
