+++ apps/dotcom/client/src/tla/components/TlaSidebar/components/sidebar-shared.ts
@@ -10,8 +10,4 @@ export const messages = defineMessages({
 	toggleSidebar: { defaultMessage: 'Toggle sidebar' },
 	accountMenu: { defaultMessage: 'Account menu' },
 	fileMenu: { defaultMessage: 'File menu' },
-	appDebugFlags: { defaultMessage: 'App debug flags' },
-	langAccented: { defaultMessage: 'i18n: Accented' },
-	langLongString: { defaultMessage: 'i18n: Long String' },
-	langHighlightMissing: { defaultMessage: 'i18n: Highlight Missing' },
 })
